+++ spec/rubocop/cop/bundler/gem_comment_spec.rb
@@ -1,0 +1,67 @@
+# frozen_string_literal: true
+
+RSpec.describe RuboCop::Cop::Bundler::GemComment, :config do
+  subject(:cop) { described_class.new(config) }
+
+  let(:cop_config) do
+    {
+      'Include' => ['**/Gemfile'],
+      'Whitelist' => ['rake']
+    }
+  end
+
+  context 'when investigating Ruby files' do
+    it 'does not register any offenses' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        gem('rubocop')
+      RUBY
+    end
+  end
+
+  context 'when investigating Gemfiles' do
+    context 'and the file is empty' do
+      it 'does not register any offenses' do
+        expect_no_offenses('', 'Gemfile')
+      end
+    end
+
+    context 'and the gem is commented' do
+      it 'does not register any offenses' do
+        expect_no_offenses(<<-RUBY.strip_indent, 'Gemfile')
+          # Style-guide enforcer.
+          gem 'rubocop'
+        RUBY
+      end
+    end
+
+    context 'and the gem is whitelisted' do
+      it 'does not register any offenses' do
+        expect_no_offenses(<<-RUBY.strip_indent, 'Gemfile')
+          gem 'rake'
+        RUBY
+      end
+    end
+
+    context 'and the file contains source and group' do
+      it 'does not register any offenses' do
+        expect_no_offenses(<<-RUBY.strip_indent, 'Gemfile')
+          source 'http://rubygems.org'
+
+          # Style-guide enforcer.
+          group :development do
+            # …
+          end
+        RUBY
+      end
+    end
+
+    context 'and a gem has no comment' do
+      it 'registers an offense' do
+        expect_offense(<<-GEM, 'Gemfile')
+          gem 'rubocop'
+          ^^^^^^^^^^^^^ Missing gem description comment.
+        GEM
+      end
+    end
+  end
+end